+++ src/palette.c
@@ -1,6 +1,6 @@
 #include "proto.h"
 
-void generate_palette (struct context * context, unsigned flags) {
+void generate_palette (struct context * context, unsigned long flags) {
   size_t count = (size_t) context -> image -> width * context -> image -> height * context -> image -> frames;
   void * palette = plum_malloc(context -> image, plum_color_buffer_size(0x100, context -> image -> color_format));
   uint8_t * indexes = plum_malloc(context -> image, count);
@@ -31,13 +31,13 @@ void remove_palette (struct context * context) {
   context -> image -> max_palette_index = 0;
 }
 
-unsigned plum_sort_palette (struct plum_image * image, unsigned flags) {
+unsigned plum_sort_palette (struct plum_image * image, unsigned long flags) {
   unsigned result = check_image_palette(image);
   if (!result) sort_palette(image, image -> color_format | (flags & PLUM_SORT_DARK_FIRST));
   return result;
 }
 
-unsigned plum_sort_palette_custom (struct plum_image * image, uint64_t (* callback) (void *, uint64_t), void * argument, unsigned flags) {
+unsigned plum_sort_palette_custom (struct plum_image * image, uint64_t (* callback) (void *, uint64_t), void * argument, unsigned long flags) {
   if (!callback) return PLUM_ERR_INVALID_ARGUMENTS;
   unsigned error = check_image_palette(image);
   if (error) return error;
@@ -62,7 +62,7 @@ unsigned plum_sort_palette_custom (struct plum_image * image, uint64_t (* callba
   return 0;
 }
 
-void sort_palette (struct plum_image * image, unsigned flags) {
+void sort_palette (struct plum_image * image, unsigned long flags) {
   uint8_t indexes[0x100];
   plum_sort_colors(image -> palette, image -> max_palette_index, flags, indexes);
   uint8_t sorted[0x100];
@@ -70,7 +70,7 @@ void sort_palette (struct plum_image * image, unsigned flags) {
   apply_sorted_palette(image, flags, sorted);
 }
 
-void apply_sorted_palette (struct plum_image * image, unsigned flags, const uint8_t * sorted) {
+void apply_sorted_palette (struct plum_image * image, unsigned long flags, const uint8_t * sorted) {
   size_t limit = (size_t) image -> width * image -> height * image -> frames;
   for (size_t p = 0; p < limit; p ++) image -> data8[p] = sorted[image -> data8[p]];
   #define sortpalette(bits) do {                                                                                         \
@@ -161,7 +161,7 @@ int plum_get_highest_palette_index (const struct plum_image * image) {
   return result;
 }
 
-int plum_convert_colors_to_indexes (uint8_t * restrict destination, const void * restrict source, void * restrict palette, size_t count, unsigned flags) {
+int plum_convert_colors_to_indexes (uint8_t * restrict destination, const void * restrict source, void * restrict palette, size_t count, unsigned long flags) {
   if (!(destination && source && palette && count)) return -PLUM_ERR_INVALID_ARGUMENTS;
   uint64_t * colors = malloc(0x800 * sizeof *colors);
   uint64_t * sorted = malloc(0x100 * sizeof *sorted);
@@ -246,7 +246,7 @@ int plum_convert_colors_to_indexes (uint8_t * restrict destination, const void *
   return result;
 }
 
-uint64_t get_color_sorting_score (uint64_t color, unsigned flags) {
+uint64_t get_color_sorting_score (uint64_t color, unsigned long flags) {
   color = plum_convert_color(color, flags, PLUM_COLOR_64 | PLUM_ALPHA_INVERT);
   uint64_t red = color & 0xffffu, green = (color >> 16) & 0xffffu, blue = (color >> 32) & 0xffffu, alpha = color >> 48;
   uint64_t luminance = red * 299 + green * 587 + blue * 114; // 26 bits
@@ -255,7 +255,8 @@ uint64_t get_color_sorting_score (uint64_t color, unsigned flags) {
   return ~((luminance << 27) | (sum << 9) | (alpha >> 7)); // total: 53 bits
 }
 
-void plum_convert_indexes_to_colors (void * restrict destination, const uint8_t * restrict source, const void * restrict palette, size_t count, unsigned flags) {
+void plum_convert_indexes_to_colors (void * restrict destination, const uint8_t * restrict source, const void * restrict palette, size_t count,
+                                     unsigned long flags) {
   if (!(destination && source && palette)) return;
   if ((flags & PLUM_COLOR_MASK) == PLUM_COLOR_16) {
     uint16_t * dp = destination;
@@ -272,7 +273,7 @@ void plum_convert_indexes_to_colors (void * restrict destination, const uint8_t 
   }
 }
 
-void plum_sort_colors (const void * restrict colors, uint8_t max_index, unsigned flags, uint8_t * restrict result) {
+void plum_sort_colors (const void * restrict colors, uint8_t max_index, unsigned long flags, uint8_t * restrict result) {
   // returns the ordered color indexes
   if (!(colors && result)) return;
   uint64_t keys[0x100]; // allocate on stack to avoid dealing with malloc() failure
